+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## 1.8.0
+
 - Feature (`@grafana/faro-web-sdk`): track `web vitals` attribution (#595).
 - Feature (`@grafana/faro-web-sdk`): set span context for navigation events (#608).
 - Feature (`@grafana/faro-react`): add helper functions to initialize React Router integration (#622).